Corrections in comm-w2c.ch.
Initial experiments with 'HAVE_GETTEXT' show that 'kpse_var_expand'
returns the expansion of '$SELFAUTOLOC' correctly, but returns the yet
to be defined '$TEXMFLOCALEFILES' AS IS (and not, as advertised in
'texk/kpathsea/variable.h', the empty string, nor the originally assumed
NULL pointer). 'strcmp'aring the string returned from 'kpse_var_expand'
to the configuration variable works as expected.

Secondly, the numeric part after flag '-d/+d' is skipped and not
interpreted as flags, i.e., we do not set numeric flags '2' and '3' from
command line option '+d32'.

git-svn-id: svn://tug.org/texlive/trunk/Build/source@49726 c570f23f-e606-0410-a88d-b1316a301751