[C/WEB] Generic solution for non-TEXMF programs.
Finally I was able to wrap my head around the 'toc-to-front' stuff!

For standard WEB and CWEB programs, place the patched '\con' exactly
between the very first '\MN' and the following '\eject'.  At the same
time, ackowledge any local definition of '\topofcontents' from the
preamble.  For this 'pdfwebtocfront.tex' must be '\input' at the very
end of the preamble, right before the first '\N'.

The four programs 'mf.web', 'tex,web', 'pdftex.web', and 'xetex.web'
define their special version of '\N' in the preamble, so the old variant
of the '\con in \N' mechanism is preserved.  Their tables-of-contents
use a bare layout anyway.

git-svn-id: svn://tug.org/texlive/trunk/Build/source@55604 c570f23f-e606-0410-a88d-b1316a301751