[CWEB] Permit indentation of preprocessing directives.
The C processor allows whitespace before and after the '#' of
preprocessing directives, e.g., some MetaPost code uses this feature.

In order to permit more liberal code layout, we only expect the '#'
to be the first character in the buffer, followed by some optional
whitespace. CWEB takes care that 'id_first' is the first non-whitespace
character in the 'buffer' (and 'id_loc' is the first character _after_
the directive name).

git-svn-id: svn://tug.org/texlive/trunk/Build/source@62084 c570f23f-e606-0410-a88d-b1316a301751