[WEB] Rewrite 'squash' to match description in section 148.
The original special case 'squash(..,1,..)' was /not/ a combination of
'app1()' and 'reduce(..,1,..)' as advertised in section 148; instead, it
was an ultra-compact variant of 'reduce'.

Moving the special case 'k==1' from 'sq' to 'red' as separate case
'k==0' makes 'squash(..,k,..)' consistent for 'k=1,2,3' (and it makes
use of the otherwise abandoned 'confusion' macro).  Curiously enough,
now there's no 'squash(..,1,..)' anymore, but CWEAVE has at least two
rules that apply this case (and it even works for 'k=4').

git-svn-id: svn://tug.org/texlive/trunk/Build/source@64063 c570f23f-e606-0410-a88d-b1316a301751